+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Initial `finishes` crate with logging and CLI skeleton
+
+### Changed
+- Bump `repo-harvest` crate version to 0.2.0
+
 ## [0.1.1] - 2025-08-17
 ### Added
 - Arch Linux setup and usage instructions in README